+++ src/openrct2/world/Park.h
@@ -65,7 +65,7 @@ namespace OpenRCT2
         void Initialise();
         void Update(const Date& date);
 
-        int32_t CalculateParkSize() const;
+        uint32_t CalculateParkSize() const;
         int32_t CalculateParkRating() const;
         money64 CalculateParkValue() const;
         money64 CalculateCompanyValue() const;
@@ -90,7 +90,7 @@ namespace OpenRCT2
 extern uint64_t gParkFlags;
 extern uint16_t gParkRating;
 extern money16 gParkEntranceFee;
-extern uint16_t gParkSize;
+extern uint32_t gParkSize;
 extern money16 gLandPrice;
 extern money16 gConstructionRightsPrice;
 
@@ -110,7 +110,7 @@ void set_forced_park_rating(int32_t rating);
 int32_t get_forced_park_rating();
 
 int32_t park_is_open();
-int32_t park_calculate_size();
+uint32_t park_calculate_size();
 
 void update_park_fences(const CoordsXY& coords);
 void update_park_fences_around_tile(const CoordsXY& coords);
